refactor(server): move lease-registry from app/ to task/leases/
app/lease-registry.ts was the largest misfiled file in the server
(1163 lines) — a dispatch-lease state machine living in the app
composition folder. The leases are task-domain state (claim → bind →
resolve lifecycle per task dispatch), not server composition.

Pure relocation: app/lease-registry.ts → task/leases/lease-registry.ts.
6 consumers updated:
  - app/index.ts re-export path
  - app/handlers/apps.handlers.ts (uses leaseRecordToWire)
  - app/types.ts (LeaseRegistry type ref)
  - app/layers.ts (defines LeaseRegistryTag + Live; uses makeLeaseRegistry)
  - app/app-host.ts (uses LeaseRegistry, LeaseVerdict types)
  - task/handlers/messages.handlers.ts (uses LeaseInvalidError + LeaseRegistry)

The file's internal ../transport/connection.js import becomes
../../transport/connection.js to follow the move; no other edits.

Smoke: tsc --noEmit clean, 29/29 server suites pass (223 tests),
pnpm -r lint green.

Follow-up: lease-registry is structured as one big file (types +
wire helpers + internal state ops + public ops + factory). At
1163 lines it fits under the eslint max-lines cap (skips blanks +
comments), but the multiple concerns inside justify a split into
~3 files (lease-types, lease-wire, lease-registry). Deferred to a
separate PR to keep this one a pure-move diff.
